+++ config/routes/api_v1_routes.rb
@@ -131,6 +131,11 @@ v1_api_routes = lambda do
     end
   end
 
+  # namespace :rsi do
+  #   resources :citizens, only: [:show], param: :handle
+  #   resources :orgs, only: %i[show], param: :sid
+  # end
+
   resource :stats, only: [] do
     get 'quick-stats' => 'stats#quick_stats'
     get 'models-per-month' => 'stats#models_per_month'