+++ lib/rsi_base_loader.rb
@@ -7,8 +7,16 @@ class RsiBaseLoader
     @base_url = options[:base_url] || Rails.application.secrets[:rsi_endpoint]
   end
 
+  private def fetch_remote(url)
+    response = Typhoeus.get(url)
+
+    AdminMailer.notify_block.deliver_later if response.code == 403
+
+    response
+  end
+
   private def strip_name(name)
-    name.gsub(/(?:AEGIS|Aegis|ANVIL|Anvil|BANU|Banu|DRAKE|Drake|ESPERIA|Esperia|KRUGER|Kruger|Kruger Intergalactic|MISC|ORIGIN|Origin|RSI|TUMBRIL|Tumbril|VANDUUL|Vanduul|Xi'an|Consolidated Outland)/, '').strip
+    name.gsub(/(?:AEGIS|Aegis|ANVIL|Anvil|BANU|Banu|Crusader|CRUSADER|DRAKE|Drake|ESPERIA|Esperia|KRUGER|Kruger|Kruger Intergalactic|MISC|ORIGIN|Origin|RSI|TUMBRIL|Tumbril|VANDUUL|Vanduul|Xi'an|Consolidated Outland)/, '').strip
   end
 
   private def nil_or_float(value)