+++ lib/rsi_roadmap_loader.rb
@@ -2,21 +2,38 @@
 
 require 'rsi_base_loader'
 
+# loader = RsiRoadmapLoader.new; ENV['RSI_LOAD_FROM_FILE'] = 'true'; loader.fetch
+
 class RsiRoadmapLoader < RsiBaseLoader
+  attr_accessor :json_file_path
+
+  def initialize(options = {})
+    super
+
+    @json_file_path = 'public/roadmap.json'
+  end
+
   def fetch
     return if roadmap_maintenance_on?
 
-    response = Typhoeus.get("#{base_url}/api/roadmap/v1/boards/1")
-    return false, nil unless response.success?
+    roadmap_data = load_roadmap_data
+
+    parse_roadmap(roadmap_data)
+  end
+
+  def load_roadmap_data
+    return JSON.parse(File.read(json_file_path))['data']['releases'] if (Rails.env.test? || ENV['CI'] || ENV['RSI_LOAD_FROM_FILE']) && File.exist?(json_file_path)
+
+    response = fetch_remote("#{base_url}/api/roadmap/v1/boards/1?#{Time.zone.now.to_i}")
+
+    return [] unless response.success?
 
     begin
       roadmap_data = JSON.parse(response.body)
-
-      items = parse_roadmap(roadmap_data)
-
-      # cleanup_changes
-
-      items
+      File.open(json_file_path, 'w') do |f|
+        f.write(roadmap_data.to_json)
+      end
+      roadmap_data['data']['releases']
     rescue JSON::ParserError => e
       Raven.capture_exception(e)
       Rails.logger.error "Roadmap Data could not be parsed: #{response.body}"
@@ -25,15 +42,18 @@ class RsiRoadmapLoader < RsiBaseLoader
   end
 
   private def roadmap_maintenance_on?
-    response = Typhoeus.get("#{base_url}/roadmap/board/1-Star-Citizen")
+    return false if Rails.env.test? || ENV['CI'] || ENV['RSI_LOAD_FROM_FILE']
+
+    response = fetch_remote("#{base_url}/roadmap/board/1-Star-Citizen?#{Time.zone.now.to_i}")
 
     !response.success?
   end
 
   # rubocop:disable Metrics/MethodLength
+  # rubocop:disable Metrics/CyclomaticComplexity
   private def parse_roadmap(data)
     roadmap_item_ids = []
-    data['data']['releases'].each do |release|
+    data.each do |release|
       release['cards'].each do |card|
         item = RoadmapItem.find_or_create_by(rsi_id: card['id']) do |new_item|
           new_item.release = release_name(new_item, release)
@@ -70,7 +90,7 @@ class RsiRoadmapLoader < RsiBaseLoader
         if item.store_image.blank?
           image_url = card['thumbnail']['urls']['source']
           image_url = "#{base_url}#{image_url}" unless image_url.starts_with?('https')
-          if image_url.present?
+          if image_url.present? && !Rails.env.test? && !ENV['CI'] && !ENV['RSI_LOAD_FROM_FILE']
             item.remote_store_image_url = image_url
             item.save
           end
@@ -87,8 +107,11 @@ class RsiRoadmapLoader < RsiBaseLoader
     RoadmapItem.where.not(id: roadmap_item_ids).find_each do |roadmap_item|
       roadmap_item.update(active: false)
     end
+
+    roadmap_item_ids
   end
   # rubocop:enable Metrics/MethodLength
+  # rubocop:enable Metrics/CyclomaticComplexity
 
   private def strip_roadmap_name(name)
     strip_name(name).gsub(/(?:Improvements|Update|Rework|Revision)/, '').strip